fix tests and make some adjustmetns in achievements

+++ SpaceFight/Assets/Kugushev/Scripts/Mission/Constants/GameplayConstants.cs
@@ -7,7 +7,7 @@
 
         public const int SoftCapArmyPower = 50;
 
-        public const int UnifiedDamage = 1;
+        public const float UnifiedDamage = 1;
         public const float SiegeRoundDelay = 0.4f;
         public const float FightRoundDelay = 0.6f;
         public const int ArmyCannonsCount = 3;
