+++ SpaceFight/Assets/Kugushev/Scripts/Mission/Utils/MissionEventsCollector.cs
@@ -16,10 +16,10 @@ namespace Kugushev.Scripts.Mission.Utils
 
         public void Cleanup() => _events.Clear();
 
-        public void PlanetCaptured(Faction newOwner, Faction previousOwner, int overpower) =>
+        public void PlanetCaptured(Faction newOwner, Faction previousOwner, float overpower) =>
             _events.Add(new MissionEvent(MissionEventType.PlanetCaptured, newOwner, previousOwner, overpower));
 
-        public void ArmyDestroyedInFight(Faction destroyer, Faction victim, int overpower) =>
+        public void ArmyDestroyedInFight(Faction destroyer, Faction victim, float overpower) =>
             _events.Add(new MissionEvent(MissionEventType.ArmyDestroyedInFight, destroyer, victim, overpower));
     }
 }
