+++ SpaceFight/Assets/Kugushev/Scripts/Mission/ValueObjects/FleetPropertiesBuilder.cs
@@ -2,7 +2,7 @@
 {
     public struct FleetPropertiesBuilder
     {
-        public int SiegeMultiplier { get; set; }
-        public int FightMultiplier { get; set; }
+        public float SiegeMultiplier { get; set; }
+        public float FightMultiplier { get; set; }
     }
 }
